+++ CUDA files/fractals/mandelbrot.cuh
@@ -3,5 +3,6 @@
 
 __global__ void fractal_rendering(
     unsigned char* pixels, size_t size_of_pixels, int width, int height,
-    double zoom_x, double zoom_y, double x_offset, double y_offset,
-    sf::Color* d_palette, int paletteSize, double maxIterations, bool* stopFlagDevice);
+    float zoom_x, float zoom_y, float x_offset, float y_offset,
+    sf::Color* d_palette, int paletteSize, float maxIterations, bool* stopFlagDevice
+);
